modules/german_sentiment/requirements.txt
hcai-discover-utils
germansentiment
